fix: cleaned, fixed and optimized layout code of navigation bars and drawers
