+++ frontend/src/components/Gallery/FabOverlay.vue
@@ -270,7 +270,7 @@ function onDownload() {
   bottom: 0;
   left: 0;
   width: 100%;
-  z-index: 1000;
+  z-index: 9999;
   pointer-events: none;
   padding-right: 8px !important;
 }
